+++ modules/client/test/src/com/worxbend/codeberg4s/AttemptParitySuite.scala
@@ -109,6 +109,7 @@ object AttemptParitySuite:
     Pair(classOf[issues.IssueSubscriptionApi], classOf[issues.IssueSubscriptionApi.Attempt]),
     Pair(classOf[pulls.PullRequestApi], classOf[pulls.PullRequestApi.Attempt]),
     Pair(classOf[pulls.PullRequestReviewApi], classOf[pulls.PullRequestReviewApi.Attempt]),
+    Pair(classOf[pulls.ReviewCommentApi], classOf[pulls.ReviewCommentApi.Attempt]),
     Pair(classOf[miscellaneous.MiscellaneousApi], classOf[miscellaneous.MiscellaneousApi.Attempt]),
     Pair(classOf[notifications.NotificationApi], classOf[notifications.NotificationApi.Attempt]),
     Pair(classOf[organizations.OrganizationApi], classOf[organizations.OrganizationApi.Attempt]),